Update README.md for Revit and .NET version support
Updated Revit version badges to reflect support for 2021-2024 and added a new badge for 2025-2026. Updated .NET framework badges to indicate support for .NET 4.8 and .NET 8. Added a badge for the last GitHub commit and included information about NBS Chorus syncing additional parameter values to Revit, with a link to the relevant support article.